+++ readme/usage.dart
@@ -5,13 +5,34 @@ const identifier = '1C709B20';
 final polar = Polar();
 
 void example() {
-  polar.heartRateStream.listen((e) => debugPrint('Heart rate: ${e.data.hr}'));
-  polar.bleSdkFeatureReadyStream.listen((e) {
-    if (e.features.contains(DeviceStreamingFeature.ecg)) {
-      polar
-          .startEcgStreaming(e.identifier)
-          .listen((e) => debugPrint('ECG data: ${e.samples}'));
-    }
-  });
   polar.connectToDevice(identifier);
+  streamWhenReady();
+}
+
+void streamWhenReady() async {
+  await polar.sdkFeatureReady.firstWhere(
+    (e) =>
+        e.identifier == identifier &&
+        e.feature == PolarSdkFeature.onlineStreaming,
+  );
+  final availabletypes =
+      await polar.getAvailableOnlineStreamDataTypes(identifier);
+
+  debugPrint('available types: $availabletypes');
+
+  if (availabletypes.contains(PolarDataType.hr)) {
+    polar
+        .startHrStreaming(identifier)
+        .listen((e) => debugPrint('HR data received'));
+  }
+  if (availabletypes.contains(PolarDataType.ecg)) {
+    polar
+        .startEcgStreaming(identifier)
+        .listen((e) => debugPrint('ECG data received'));
+  }
+  if (availabletypes.contains(PolarDataType.acc)) {
+    polar
+        .startAccStreaming(identifier)
+        .listen((e) => debugPrint('ACC data received'));
+  }
 }